Make printing workboards work OK
Summary:
Fixes T7844. This isn't going to win any awards, but it has all the information.

Mostly, we don't (or shouldn't, at least?) need the global `static` stuff anymore because we dropped the top-level custom scrollbar.

Test Plan: Printed a PDF of a workboard, got all the cards in the output.

Reviewers: chad

Reviewed By: chad

Maniphest Tasks: T7844

Differential Revision: https://secure.phabricator.com/D15205